+++ master-FE/src/components/Admin/ViewStaffList.jsx
@@ -3,6 +3,9 @@ import { deleteUserApi, getAllUserApi } from "../../services/apiCall";
 import Table from "../common/Table";
 import toast from "react-hot-toast";
 import { useNavigate } from "react-router-dom";
+import { confirmAlert } from "react-confirm-alert";
+import "react-confirm-alert/src/react-confirm-alert.css";
+import "../../confimbox.css";
 
 function ViewStaffList() {
   const [staffData, setStaffData] = useState([]);
@@ -48,17 +51,37 @@ function ViewStaffList() {
   };
 
   const handleDelete = async (row) => {
-    try {
-      const response = await deleteUserApi(row.id, "staff");
-      staffData.map((student) =>
-        student.id !== row.id
-          ? student
-          : setStaffData(staffData.filter((student) => student.id !== row.id))
-      );
-      toast.success(response.data.message);
-    } catch (error) {
-      toast.error(error.response.data.message);
-    }
+    confirmAlert({
+      customUI: ({ onClose }) => {
+        return (
+          <div className="custom-alert-overlay">
+            <div className="custom-alert">
+              <h1>Confirm to delete</h1>
+              <p>Are you sure you want to delete this staff?</p>
+              <button
+                onClick={async () => {
+                  try {
+                    const response = await deleteUserApi(row.id, "staff");
+                    setStaffData(
+                      staffData.filter((staff) => staff.id !== row.id)
+                    );
+                    toast.success(response.data.message);
+                  } catch (error) {
+                    toast.error(error.response.data.message);
+                  }
+                  onClose();
+                }}
+              >
+                Yes
+              </button>
+              <button onClick={onClose} className="cancel">
+                No
+              </button>
+            </div>
+          </div>
+        );
+      },
+    });
   };
 
   const handleSort = (key) => {
